clean code of auth & apart

+++ app/Http/Controllers/AuthController.php
@@ -1,42 +1,28 @@
 <?php
 namespace App\Http\Controllers;
+
+use App\Http\Requests\AdminloginRequest;
+use App\Http\Requests\OtpRequestue;
+use App\Http\Requests\SigninRequest;
+use App\Http\Requests\SignupRequest;
 use Illuminate\Http\Request;
 use App\Models\User;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Storage;
-use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Http;
 use Carbon\Carbon;
 class AuthController extends Controller
 {
-    public function signup(Request $request)
+// sign up the customer to the first time and choose his rule :tenant ,owner
+    public function signup(SignupRequest $request)
 {
     try {
-        $validator = Validator::make($request->all(), [
-            'mobile_phone' => 'required|unique:users|regex:/^[0-9]{10,15}$/',
-            'password' => 'required|min:6',
-            'role' => 'required|in:tenant,owner',
-            'first_name' => 'required|string|max:100',
-            'last_name' => 'required|string|max:100',
-            'date_of_birth' => 'required|date|before:today',
-            'profile_image' => 'required|image|mimes:jpeg,png,jpg|max:2048',
-            'id_card_image' => 'required|image|mimes:jpeg,png,jpg|max:2048',
-        ]);
-
-        if ($validator->fails()) {
-            return response()->json([
-                'status' => 'error',
-                'message' => 'data is invalid',
-                'errors' => $validator->errors()
-            ], 422);
-        }
         $profilePath = $request->file('profile_image')->storeAs(
             'profiles',
             uniqid().'_'.$request->file('profile_image')->getClientOriginalName(),
             'public'
         );
-
         $idCardPath = $request->file('id_card_image')->storeAs(
             'id_cards',
             uniqid().'_'.$request->file('id_card_image')->getClientOriginalName(),
@@ -66,18 +52,30 @@ class AuthController extends Controller
 
         Log::error($e);
         return response()->json([
-            'status' => 'error',
-            'message' => 'wrong in the sign up , try again'
+            'status' => 'Error',
+            'message' => 'Error in the sign up , try again'
         ], 500);
     }
 }
-public function adminLogin(Request $request)
+// operation of send the otp code
+public function sendOTP(User $user){
+        $otp = rand(100000, 999999);
+        $user->update([
+            'otp_code' => $otp,
+            'otp_expires_at' => Carbon::now()->addMinutes(5),
+        ]);
+        $instanceId = env('ULTRAMSG_INSTANCE_ID');
+        $token = env('ULTRAMSG_TOKEN');
+        Http::post("https://api.ultramsg.com/$instanceId/messages/chat", [
+            'token' => $token,
+            'to'    => $user->mobile_phone,
+            'body'  => "your otp code is :$otp",
+        ]);
+}
+//admin login and send the otp code to his whatsapp
+public function adminLogin(AdminloginRequest $request)
 {
     try {
-        $request->validate([
-            'mobile_phone' => 'required',
-            'password' => 'required'
-        ]);
         $user = User::where('mobile_phone', $request->mobile_phone)->first();
         if (!$user || !Hash::check($request->password, $user->password)) {
             return response()->json(['message' => 'Invalid credentials'], 401);
@@ -85,115 +83,69 @@ public function adminLogin(Request $request)
         if (!$user->isAdmin()) {
             return response()->json(['message' => 'Access denied. Admin only.'], 403);
         }
-        $otp = rand(100000, 999999);
-        $user->update([
-            'otp_code' => $otp,
-            'otp_expires_at' => Carbon::now()->addMinutes(5),
-        ]);
-        $instanceId = env('ULTRAMSG_INSTANCE_ID');
-        $token = env('ULTRAMSG_TOKEN');
-
-        Http::post("https://api.ultramsg.com/$instanceId/messages/chat", [
-            'token' => $token,
-            'to'    => $user->mobile_phone,
-            'body'  => "your pin code is :$otp",
-        ]);
-
-        return response()->json(['message' => 'send it in whatsapp']);
+        $this->sendOTP($user);
+        return response()->json(['message' => 'send the code to whatsapp']);
     } catch (\Exception $e) {
         Log::error($e);
-        return response()->json(['message' => 'Error in admin login'], 500);
+        return response()->json(['message' => 'Error in login , try again'], 500);
     }
 }
-public function signin(Request $request)
+// sign in after admin approved and send the otp code
+public function signin(SigninRequest $request)
 {
     try {
-        $request->validate([
-            'mobile_phone' => 'required',
-            'password' => 'required'
-        ]);
-
         $user = User::where('mobile_phone', $request->mobile_phone)->first();
-
         if (!$user || !Hash::check($request->password, $user->password)) {
             return response()->json(['message' => 'Invalid credentials'], 401);
         }
-
         if (!$user->is_approved) {
-            return response()->json(['message' => 'pending admin approval'], 403);
+            return response()->json(['message' => 'pending admin approved'], 403);
         }
-        $otp = rand(100000, 999999);
-        $user->update([
-            'otp_code' => $otp,
-            'otp_expires_at' => Carbon::now()->addMinutes(5),
-        ]);
-        $instanceId = env('ULTRAMSG_INSTANCE_ID');
-        $token = env('ULTRAMSG_TOKEN');
-
-        Http::post("https://api.ultramsg.com/$instanceId/messages/chat", [
-            'token' => $token,
-            'to'    => $user->mobile_phone,
-            'body'  => "your pin code is :$otp",
-        ]);
-
-        return response()->json(['message' => 'send it in whatsapp']);
+        $this->sendOTP($user);
+        return response()->json(['message' => 'send the code to whatsapp']);
     } catch (\Exception $e) {
         Log::error($e);
-        return response()->json(['message' => 'wrong in sign in , try again'], 500);
+        return response()->json(['message' => 'Error in sign in , try again'], 500);
     }
 }
-
+//logout the user from  the current token
 public function logout(Request $request)
     {
         try {
-            $request->user()->currentAccessToken()->delete();
-            
-            return response()->json([
-                'message' => 'logout successfully'
-            ]);
+            if ($request->user()->currentAccessToken()) {
+                $request->user()->currentAccessToken()->delete();
+            }
+            return response()->json(['message' => 'logout successfully']);
         } catch (\Exception $e) {
             Log::error($e);
-            return response()->json([
-            'message' => 'wrong in logout ,try again'
-            ], 500);
+            return response()->json(['message' => 'Error in logout, try again'], 500);
         }
     }
-
-public function verifyOtp(Request $request)
+//verify fron otp that send it to the user in whatsapp
+public function verifyOtp(OtpRequestue $request)
 {
-    $request->validate([
-        'mobile_phone' => 'required',
-        'otp_code' => 'required'
-    ]);
-
     $user = User::where('mobile_phone', $request->mobile_phone)->first();
-
     if (!$user) {
         return response()->json(['message' => 'User not found'], 404);
     }
-
     if ((string)$user->otp_code !== (string)$request->otp_code) {
         return response()->json(['message' => 'Invalid OTP'], 401);
     }
-
     if (Carbon::now()->greaterThan($user->otp_expires_at)) {
         return response()->json(['message' => 'OTP expired'], 403);
     }
-
     $user->update([
         'otp_code' => null,
         'otp_expires_at' => null,
     ]);
-
     $user->tokens()->delete();
     $token = $user->createToken('auth_token')->plainTextToken;
-
     return response()->json([
         'token' => $token,
         'user' => $user,
     ]);
 }
-
+//user with pending status waiting to approve from admin
     public function pendingUsers(Request $request)
     {
         if (!$request->user()->isAdmin()) {
@@ -206,7 +158,7 @@ public function verifyOtp(Request $request)
             'message'=> 'Pending users retrieved successfully.'
         ]);
     }
-
+// admin approved to the user
     public function approveUser(Request $request, User $user)
     {
         if (!$request->user()->isAdmin()) {
@@ -219,30 +171,26 @@ public function verifyOtp(Request $request)
             'data'    => $user
         ]);
     }
-
+//admin reject the user
     public function rejectUser(Request $request, User $user)
     {
         if (!$request->user()->isAdmin()) {
             return response()->json(['message' => 'Unauthorized'], 403);
         }
-
         $user->update(['is_approved' => false]);
-
         return response()->json([
             'status'  => 'success',
             'message' => 'User rejected successfully.',
             'data'    => $user
         ]);
     }
-
+//admin delete the user
     public function deleteUser(Request $request, User $user)
     {
         if (!$request->user()->isAdmin()) {
             return response()->json(['message' => 'Unauthorized'], 403);
         }
-
         $user->delete();
-
         return response()->json([
             'status'  => 'success',
             'message' => 'User deleted successfully.'
